Extract Expression to its own file
