+++ packages/alfa-css/tsconfig.json
@@ -15,6 +15,9 @@
     "src/value/box.ts",
     "src/value/calculation.ts",
     "src/value/calculation/calculation.ts",
+    "src/value/calculation/expression.ts",
+    "src/value/calculation/kind.ts",
+    "src/value/calculation/value.ts",
     "src/value/color.ts",
     "src/value/color/current.ts",
     "src/value/color/hex.ts",
